feat : 2023_11_01 알림처리 관련 기능 수정
알림 받을 게시글이나 댓글이 삭제 되었을 경우에...
DB에서 알림 지우고 프론트로 쏘지 않기
일단... 공지사항만..!!

+++ src/main/java/com/kh/Freepets/service/board/notice/NoticeService.java
@@ -74,7 +74,8 @@ public class NoticeService
     public Notice delete(int noticeCode)
     {
         Notice target = dao.findById(noticeCode).orElse(null);
-        dao.delete(target);
+        if (target != null)
+            dao.delete(target);
         return target;
     }
